requirements.txt
# /********************************************************************************
# * Copyright (c) 2021 Contributors to the Eclipse Foundation
# *
# * See the NOTICE file(s) distributed with this work for additional
# * information regarding copyright ownership.
# *
# * This program and the accompanying materials are made available under the
# * terms of the Eclipse Public License 2.0 which is available at
# * http://www.eclipse.org/legal/epl-2.0
# *
# * SPDX-License-Identifier: EPL-2.0
# ********************************************************************************/

grpcio>=1.26.0
grpcio-tools>=1.26.0
protobuf==3.17.3
cloudevents >= 1.0.0
opencensus == 0.8.0
opencensus-ext-grpc == 0.7.2
dapr-ext-grpc-dev >= 1.4.0rc1.dev
dapr-dev >= 1.4.0rc1.dev